fix(budget): Use budget type over budget id
Fixes the exception thrown when getting the budget summary.

Budgets were never saving the transaction type on creation or update. This caused the query to fail. To fix this, the transaction type table was fully removed and replaced with an enum that defined it on the budget category level rather than on the budget or transaction level. Transactions are now evaluated to be expense or income based on whether their value is negative or positive.

BREAKING CHANGE